Move plex guid providers to own namespace
